Create components SubscriptionFree, SubscriptionBusiness, SubscriptionPremium
